feat(user): add guest mode with local task storage
Allow users to continue as a guest without needing to log in. Guest users can now create and manage tasks, which are saved locally in the browser. This provides a seamless experience for users who want
to quickly use the app without signing up.

fix(task editing): resolve "Remove Category" button issue

Fixed an issue where the "Remove Category" button was not performing any action when clicked. Updated the button's event handler to ensure that categories are correctly removed.

+++ src/hooks/useDeleteTask.js
@@ -8,7 +8,10 @@ import useTaskStore from "../store/taskStore";
 function useDeleteTask() {
     const showToast = useShowToast();
 
-    const authUser = useAuthStore((state) => state.user);
+    const { authUser, isLoggedIn } = useAuthStore((state) => ({
+        authUser: state.user,
+        isLoggedIn: state.isLoggedIn,
+    }));
     const { tasks, setTasks, deleteTasks, sortTasks } = useTaskStore();
 
     const [isDeleting, setIsDeleting] = useState(false);
@@ -19,19 +22,49 @@ function useDeleteTask() {
         async (tasksToDelete) => {
             if (isDeleting) return;
 
+            setIsDeleting(true);
+
             try {
-                setIsDeleting(true);
-                const userRef = doc(firestore, "users", authUser.uid);
+                if (isLoggedIn) {
+                    const userRef = doc(firestore, "users", authUser.uid);
 
-                // Delete each task and update the user's tasks array
-                const deletePromises = tasksToDelete.map(async (task) => {
-                    await deleteDoc(doc(firestore, "tasks", task.id));
-                    await updateDoc(userRef, {
-                        tasks: arrayRemove(task.id),
+                    // Delete each task and update the user's tasks array
+                    const deletePromises = tasksToDelete.map(async (task) => {
+                        await deleteDoc(doc(firestore, "tasks", task.id));
+                        await updateDoc(userRef, {
+                            tasks: arrayRemove(task.id),
+                        });
                     });
-                });
 
-                await Promise.all(deletePromises);
+                    await Promise.all(deletePromises);
+                } else {
+                    // If user is not logged in, delete locally
+                    const storedTasks = JSON.parse(
+                        localStorage.getItem("tasks")
+                    );
+                    if (storedTasks) {
+                        setTasks(
+                            storedTasks.filter(
+                                (task) =>
+                                    !tasksToDelete.some(
+                                        (delTask) => delTask.id === task.id
+                                    )
+                            )
+                        );
+                    }
+
+                    localStorage.setItem(
+                        "tasks",
+                        JSON.stringify(
+                            tasks.filter(
+                                (task) =>
+                                    !tasksToDelete.some(
+                                        (delTask) => delTask.id === task.id
+                                    )
+                            )
+                        )
+                    );
+                }
 
                 // Update the state to remove the deleted tasks
                 setTasks(
